requirements.txt
requests>=1.1.0
redis>=2.10.3
Babel>=1.3
netaddr>=0.7.12
six>=1.6.1
python-swiftclient>=2.2.0
oslo.utils>=0.2.0
PrettyTable>=0.7,<0.8
websocket-client>=0.16.0